Added endpoint group data file reading.
git-svn-id: svn://biometris-nas.wurnet.nl/AMIGA@307 3ae013c2-4042-2242-b34d-985be216b4aa

+++ AmigaPowerAnalysis/AmigaPowerAnalysis.Core/Core/Data/EndpointGroupDTO.cs
@@ -8,7 +8,7 @@ namespace AmigaPowerAnalysis.Core.Data {
         public double LocLower { get; set; }
         public double LocUpper { get; set; }
         public DistributionType DistributionType { get; set; }
-        public double BinomialTotal { get; set; }
+        public int BinomialTotal { get; set; }
         public double PowerLawPower { get; set; }
         public double MuComparator { get; set; }
         public double CvComparator { get; set; }
